minor changes added
