+++ go.mod
@@ -2,9 +2,9 @@ module github.com/kailt/imageresizer
 
 require (
 	github.com/aws/aws-sdk-go v1.15.59
-	github.com/fsnotify/fsnotify v1.4.7
 	github.com/cloudflare/tableflip v0.0.0-20181019105324-78281f93d075
 	github.com/djherbis/atime v1.0.0
+	github.com/fsnotify/fsnotify v1.4.7
 	github.com/gorilla/mux v1.6.2
 	github.com/pkg/errors v0.8.0 // indirect
 	github.com/rcrowley/go-metrics v0.0.0-20181016184325-3113b8401b8a
